requirements.txt
shiny==0.6.0
pinecone-client[grpc]==2.2.4
python-dotenv==1.0.0